Create 21-Longest Decreasing Subsequence
Problem statement
You are given an array/list ARR consisting of N integers. Your task is to find the length of the longest decreasing subsequence.

(https://www.naukri.com/code360/problems/longest-decreasing-subsequence_800300?leftPanelTabValue=PROBLEM)

A subsequence is a sequence of numbers obtained by deleting zero or more elements from the array/list, keeping the relative positions of elements same as it was in the initial sequence. A decreasing subsequence is a subsequence in which every element is strictly less than the previous number.

Note:

There can be more than one subsequences with the longest length.
For example:-
For the given array [5, 0, 3, 2, 9], the longest decreasing subsequence is of length 3, i.e. [5, 3, 2]
Note:
Try to solve the problem in O(N log N) time complexity.

Constraints:
1 <= T <= 50 
1 <= N <= 5000
1 <= ARR[i] <= 10^9

Time Limit: 1 sec
Sample Input 1:
 3
 5 
 5 1 3 2 4
 5 
 6 8 9 2 1
 5 
 1 2 3 4 5 
Sample Output 1:
 3
 3
 1
Explanation for sample input 1:
For the first test case, the longest decreasing subsequence is [5, 3, 2] of length 3.

For the second test case, the longest decreasing subsequence is [9, 2, 1] of length. Note [8,2,1] and [6,2,1] are of length 3 too.

For the third test case, the longest decreasing subsequence is [1] of length 1. Note [2], [3], [4] and [5] are of length 1 too.
Sample Input 2:
 2
 4 
 63 22 56 94 
 2 
 83 80
Sample Output 2:
 2
 2